Move SelfConsistentEffectiveMedium to _mesh_agnostic.py
