+++ simpeg/maps/_base.py
@@ -5,26 +5,20 @@ Base and general map classes.
 from __future__ import annotations  # needed to use type operands in Python 3.8
 
 from collections import namedtuple
-import warnings
 import discretize
 import numpy as np
 import scipy.sparse as sp
 from scipy.sparse import csr_matrix as csr
 from discretize.tests import check_derivative
 from discretize.utils import (
-    rotation_matrix_from_normals,
     Zero,
     Identity,
-    sdiag,
 )
 
 from ..utils import (
     mat_utils,
     validate_type,
     validate_ndarray_with_shape,
-    validate_float,
-    validate_direction,
-    validate_integer,
     validate_list_of_types,
 )
 from ..typing import RandomSeed
@@ -1180,511 +1174,3 @@ class Wires(object):
             Number of parameters that the mapping acts on.
         """
         return self._nP
-
-
-class SelfConsistentEffectiveMedium(IdentityMap):
-    r"""
-        Two phase self-consistent effective medium theory mapping for
-        ellipsoidal inclusions. The inversion model is the concentration
-        (volume fraction) of the phase 2 material.
-
-        The inversion model is :math:`\varphi`. We solve for :math:`\sigma`
-        given :math:`\sigma_0`, :math:`\sigma_1` and :math:`\varphi` . Each of
-        the following are implicit expressions of the effective conductivity.
-        They are solved using a fixed point iteration.
-
-        **Spherical Inclusions**
-
-        If the shape of the inclusions are spheres, we use
-
-        .. math::
-
-            \sum_{j=1}^N (\sigma^* - \sigma_j)R^{j} = 0
-
-        where :math:`j=[1,N]` is the each material phase, and N is the number
-        of phases. Currently, the implementation is only set up for 2 phase
-        materials, so we solve
-
-        .. math::
-
-            (1-\\varphi)(\sigma - \sigma_0)R^{(0)} + \varphi(\sigma - \sigma_1)R^{(1)} = 0.
-
-        Where :math:`R^{(j)}` is given by
-
-        .. math::
-
-            R^{(j)} = \left[1 + \frac{1}{3}\frac{\sigma_j - \sigma}{\sigma} \right]^{-1}.
-
-        **Ellipsoids**
-
-        .. todo::
-
-            Aligned Ellipsoids have not yet been implemented, only randomly
-            oriented ellipsoids
-
-        If the inclusions are aligned ellipsoids, we solve
-
-        .. math::
-
-            \sum_{j=1}^N \varphi_j (\Sigma^* - \sigma_j\mathbf{I}) \mathbf{R}^{j, *} = 0
-
-        where
-
-        .. math::
-
-            \mathbf{R}^{(j, *)} = \left[ \mathbf{I} + \mathbf{A}_j {\Sigma^{*}}^{-1}(\sigma_j \mathbf{I} - \Sigma^*) \\right]^{-1}
-
-        and the depolarization tensor :math:`\mathbf{A}_j` is given by
-
-        .. math::
-
-            \mathbf{A}^* = \left[\begin{array}{ccc}
-                Q & 0 & 0 \\
-                0 & Q & 0 \\
-                0 & 0 & 1-2Q
-            \end{array}\right]
-
-        for a spheroid aligned along the z-axis. For an oblate spheroid
-        (:math:`\alpha < 1`, pancake-like)
-
-        .. math::
-
-            Q = \frac{1}{2}\left(
-                1 + \frac{1}{\alpha^2 - 1} \left[
-                    1 - \frac{1}{\chi}\tan^{-1}(\chi)
-                \right]
-            \right)
-
-        where
-
-        .. math::
-
-            \chi = \sqrt{\frac{1}{\alpha^2} - 1}
-
-
-        For reference, see
-        `Torquato (2002), Random Heterogeneous Materials <https://link.springer.com/book/10.1007/978-1-4757-6355-3>`_
-
-
-    """
-
-    def __init__(
-        self,
-        mesh=None,
-        nP=None,
-        sigma0=None,
-        sigma1=None,
-        alpha0=1.0,
-        alpha1=1.0,
-        orientation0="z",
-        orientation1="z",
-        random=True,
-        rel_tol=1e-3,
-        maxIter=50,
-        **kwargs,
-    ):
-        self._sigstart = None
-        self.sigma0 = sigma0
-        self.sigma1 = sigma1
-        self.alpha0 = alpha0
-        self.alpha1 = alpha1
-        self.orientation0 = orientation0
-        self.orientation1 = orientation1
-        self.random = random
-        self.rel_tol = rel_tol
-        self.maxIter = maxIter
-        super(SelfConsistentEffectiveMedium, self).__init__(mesh, nP, **kwargs)
-
-    @property
-    def sigma0(self):
-        """Physical property value for phase-0 material.
-
-        Returns
-        -------
-        float
-        """
-        return self._sigma0
-
-    @sigma0.setter
-    def sigma0(self, value):
-        self._sigma0 = validate_float("sigma0", value, min_val=0.0)
-
-    @property
-    def sigma1(self):
-        """Physical property value for phase-1 material.
-
-        Returns
-        -------
-        float
-        """
-        return self._sigma1
-
-    @sigma1.setter
-    def sigma1(self, value):
-        self._sigma1 = validate_float("sigma1", value, min_val=0.0)
-
-    @property
-    def alpha0(self):
-        """Aspect ratio of the phase-0 ellipsoids.
-
-        Returns
-        -------
-        float
-        """
-        return self._alpha0
-
-    @alpha0.setter
-    def alpha0(self, value):
-        self._alpha0 = validate_float("alpha0", value, min_val=0.0)
-
-    @property
-    def alpha1(self):
-        """Aspect ratio of the phase-1 ellipsoids.
-
-        Returns
-        -------
-        float
-        """
-        return self._alpha1
-
-    @alpha1.setter
-    def alpha1(self, value):
-        self._alpha1 = validate_float("alpha1", value, min_val=0.0)
-
-    @property
-    def orientation0(self):
-        """Orientation of the phase-0 inclusions.
-
-        Returns
-        -------
-        numpy.ndarray
-        """
-        return self._orientation0
-
-    @orientation0.setter
-    def orientation0(self, value):
-        self._orientation0 = validate_direction("orientation0", value, dim=3)
-
-    @property
-    def orientation1(self):
-        """Orientation of the phase-0 inclusions.
-
-        Returns
-        -------
-        numpy.ndarray
-        """
-        return self._orientation1
-
-    @orientation1.setter
-    def orientation1(self, value):
-        self._orientation1 = validate_direction("orientation1", value, dim=3)
-
-    @property
-    def random(self):
-        """Are the inclusions randomly oriented (True) or preferentially aligned (False)?
-
-        Returns
-        -------
-        bool
-        """
-        return self._random
-
-    @random.setter
-    def random(self, value):
-        self._random = validate_type("random", value, bool)
-
-    @property
-    def rel_tol(self):
-        """relative tolerance for convergence for the fixed-point iteration.
-
-        Returns
-        -------
-        float
-        """
-        return self._rel_tol
-
-    @rel_tol.setter
-    def rel_tol(self, value):
-        self._rel_tol = validate_float(
-            "rel_tol", value, min_val=0.0, inclusive_min=False
-        )
-
-    @property
-    def maxIter(self):
-        """Maximum number of iterations for the fixed point iteration calculation.
-
-        Returns
-        -------
-        int
-        """
-        return self._maxIter
-
-    @maxIter.setter
-    def maxIter(self, value):
-        self._maxIter = validate_integer("maxIter", value, min_val=0)
-
-    @property
-    def tol(self):
-        """
-        absolute tolerance for the convergence of the fixed point iteration
-        calc
-        """
-        if getattr(self, "_tol", None) is None:
-            self._tol = self.rel_tol * min(self.sigma0, self.sigma1)
-        return self._tol
-
-    @property
-    def sigstart(self):
-        """
-        first guess for sigma
-        """
-        return self._sigstart
-
-    @sigstart.setter
-    def sigstart(self, value):
-        if value is not None:
-            value = validate_float("sigstart", value)
-        self._sigstart = value
-
-    def wiener_bounds(self, phi1):
-        """Define Wenner Conductivity Bounds
-
-        See Torquato, 2002
-        """
-        phi0 = 1.0 - phi1
-        sigWup = phi0 * self.sigma0 + phi1 * self.sigma1
-        sigWlo = 1.0 / (phi0 / self.sigma0 + phi1 / self.sigma1)
-        W = np.array([sigWlo, sigWup])
-
-        return W
-
-    def hashin_shtrikman_bounds(self, phi1):
-        """Hashin Shtrikman bounds
-
-        See Torquato, 2002
-        """
-        # TODO: this should probably exsist on its own as a util
-
-        phi0 = 1.0 - phi1
-        sigWu = self.wiener_bounds(phi1)[1]
-        sig_tilde = phi0 * self.sigma1 + phi1 * self.sigma0
-
-        sigma_min = np.min([self.sigma0, self.sigma1])
-        sigma_max = np.max([self.sigma0, self.sigma1])
-
-        sigHSlo = sigWu - (
-            (phi0 * phi1 * (self.sigma0 - self.sigma1) ** 2)
-            / (sig_tilde + 2 * sigma_max)
-        )
-        sigHSup = sigWu - (
-            (phi0 * phi1 * (self.sigma0 - self.sigma1) ** 2)
-            / (sig_tilde + 2 * sigma_min)
-        )
-
-        return np.array([sigHSlo, sigHSup])
-
-    def hashin_shtrikman_bounds_anisotropic(self, phi1):
-        """Hashin Shtrikman bounds for anisotropic media
-
-        See Torquato, 2002
-        """
-        phi0 = 1.0 - phi1
-        sigWu = self.wiener_bounds(phi1)[1]
-
-        sigma_min = np.min([self.sigma0, self.sigma1])
-        sigma_max = np.max([self.sigma0, self.sigma1])
-
-        phi_min = phi0 if self.sigma1 > self.sigma0 else phi1
-        phi_max = phi1 if self.sigma1 > self.sigma0 else phi0
-
-        amax = (
-            -phi0
-            * phi1
-            * self.getA(
-                self.alpha1 if self.sigma1 > self.sigma0 else self.alpha0,
-                self.orientation1 if self.sigma1 > self.sigma0 else self.orientation0,
-            )
-        )
-        I = np.eye(3)
-
-        sigHSlo = sigWu * I + (
-            (sigma_min - sigma_max) ** 2
-            * amax
-            * np.linalg.inv(sigma_min * I + (sigma_min - sigma_max) / phi_max * amax)
-        )
-        sigHSup = sigWu * I + (
-            (sigma_max - sigma_min) ** 2
-            * amax
-            * np.linalg.inv(sigma_max * I + (sigma_max - sigma_min) / phi_min * amax)
-        )
-
-        return [sigHSlo, sigHSup]
-
-    def getQ(self, alpha):
-        """Geometric factor in the depolarization tensor"""
-        if alpha < 1.0:  # oblate spheroid
-            chi = np.sqrt((1.0 / alpha**2.0) - 1)
-            return (
-                1.0 / 2.0 * (1 + 1.0 / (alpha**2.0 - 1) * (1.0 - np.arctan(chi) / chi))
-            )
-        elif alpha > 1.0:  # prolate spheroid
-            chi = np.sqrt(1 - (1.0 / alpha**2.0))
-            return (
-                1.0
-                / 2.0
-                * (
-                    1
-                    + 1.0
-                    / (alpha**2.0 - 1)
-                    * (1.0 - 1.0 / (2.0 * chi) * np.log((1 + chi) / (1 - chi)))
-                )
-            )
-        elif alpha == 1:  # sphere
-            return 1.0 / 3.0
-
-    def getA(self, alpha, orientation):
-        """Depolarization tensor"""
-        Q = self.getQ(alpha)
-        A = np.diag([Q, Q, 1 - 2 * Q])
-        R = rotation_matrix_from_normals(np.r_[0.0, 0.0, 1.0], orientation)
-        return (R.T).dot(A).dot(R)
-
-    def getR(self, sj, se, alpha, orientation=None):
-        """Electric field concentration tensor"""
-        if self.random is True:  # isotropic
-            if alpha == 1.0:
-                return 3.0 * se / (2.0 * se + sj)
-            Q = self.getQ(alpha)
-            return (
-                se
-                / 3.0
-                * (2.0 / (se + Q * (sj - se)) + 1.0 / (sj - 2.0 * Q * (sj - se)))
-            )
-        else:  # anisotropic
-            if orientation is None:
-                raise Exception("orientation must be provided if random=False")
-            I = np.eye(3)
-            seinv = np.linalg.inv(se)
-            Rinv = I + self.getA(alpha, orientation) * seinv * (sj * I - se)
-            return np.linalg.inv(Rinv)
-
-    def getdR(self, sj, se, alpha, orientation=None):
-        """
-        Derivative of the electric field concentration tensor with respect
-        to the concentration of the second phase material.
-        """
-        if self.random is True:
-            if alpha == 1.0:
-                return 3.0 / (2.0 * se + sj) - 6.0 * se / (2.0 * se + sj) ** 2
-            Q = self.getQ(alpha)
-            return (
-                1
-                / 3
-                * (
-                    2.0 / (se + Q * (sj - se))
-                    + 1.0 / (sj - 2.0 * Q * (sj - se))
-                    + se
-                    * (
-                        -2 * (1 - Q) / (se + Q * (sj - se)) ** 2
-                        - 2 * Q / (sj - 2.0 * Q * (sj - se)) ** 2
-                    )
-                )
-            )
-        else:
-            if orientation is None:
-                raise Exception("orientation must be provided if random=False")
-            raise NotImplementedError
-
-    def _sc2phaseEMTSpheroidstransform(self, phi1):
-        """
-        Self Consistent Effective Medium Theory Model Transform,
-        alpha = aspect ratio (c/a <= 1)
-        """
-
-        if not (np.all(0 <= phi1) and np.all(phi1 <= 1)):
-            warnings.warn("there are phis outside bounds of 0 and 1", stacklevel=2)
-            phi1 = np.median(np.c_[phi1 * 0, phi1, phi1 * 0 + 1.0])
-
-        phi0 = 1.0 - phi1
-
-        # starting guess
-        if self.sigstart is None:
-            sige1 = np.mean(self.wiener_bounds(phi1))
-        else:
-            sige1 = self.sigstart
-
-        if self.random is False:
-            sige1 = sige1 * np.eye(3)
-
-        for _ in range(self.maxIter):
-            R0 = self.getR(self.sigma0, sige1, self.alpha0, self.orientation0)
-            R1 = self.getR(self.sigma1, sige1, self.alpha1, self.orientation1)
-
-            den = phi0 * R0 + phi1 * R1
-            num = phi0 * self.sigma0 * R0 + phi1 * self.sigma1 * R1
-
-            if self.random is True:
-                sige2 = num / den
-                relerr = np.abs(sige2 - sige1)
-            else:
-                sige2 = num * np.linalg.inv(den)
-                relerr = np.linalg.norm(np.abs(sige2 - sige1).flatten(), np.inf)
-
-            if np.all(relerr <= self.tol):
-                if self.sigstart is None:
-                    self._sigstart = (
-                        sige2  # store as a starting point for the next time around
-                    )
-                return sige2
-
-            sige1 = sige2
-        # TODO: make this a proper warning, and output relevant info (sigma0, sigma1, phi, sigstart, and relerr)
-        warnings.warn("Maximum number of iterations reached", stacklevel=2)
-
-        return sige2
-
-    def _sc2phaseEMTSpheroidsinversetransform(self, sige):
-        R0 = self.getR(self.sigma0, sige, self.alpha0, self.orientation0)
-        R1 = self.getR(self.sigma1, sige, self.alpha1, self.orientation1)
-
-        num = -(self.sigma0 - sige) * R0
-        den = (self.sigma1 - sige) * R1 - (self.sigma0 - sige) * R0
-
-        return num / den
-
-    def _sc2phaseEMTSpheroidstransformDeriv(self, sige, phi1):
-        phi0 = 1.0 - phi1
-
-        R0 = self.getR(self.sigma0, sige, self.alpha0, self.orientation0)
-        R1 = self.getR(self.sigma1, sige, self.alpha1, self.orientation1)
-
-        dR0 = self.getdR(self.sigma0, sige, self.alpha0, self.orientation0)
-        dR1 = self.getdR(self.sigma1, sige, self.alpha1, self.orientation1)
-
-        num = (sige - self.sigma0) * R0 - (sige - self.sigma1) * R1
-        den = phi0 * (R0 + (sige - self.sigma0) * dR0) + phi1 * (
-            R1 + (sige - self.sigma1) * dR1
-        )
-
-        return sdiag(num / den)
-
-    def _transform(self, m):
-        return self._sc2phaseEMTSpheroidstransform(m)
-
-    def deriv(self, m):
-        """
-        Derivative of the effective conductivity with respect to the
-        volume fraction of phase 2 material
-        """
-        sige = self._transform(m)
-        return self._sc2phaseEMTSpheroidstransformDeriv(sige, m)
-
-    def inverse(self, sige):
-        """
-        Compute the concentration given the effective conductivity
-        """
-        return self._sc2phaseEMTSpheroidsinversetransform(sige)
-
-    @property
-    def is_linear(self):
-        return False
